Update project to version 4.3.232, see the changelogs for more infos.

+++ tools/Storage/prepare_update_on_sd.bat
@@ -691,11 +691,19 @@ IF NOT "%temp_volume_format%"=="FAT32" (
 IF NOT "%cancel_script%"=="" set cancel_script=%cancel_script:~0,1%
 call "tools\Storage\functions\modify_yes_no_always_never_vars.bat" "cancel_script" "o/n_choice"
 IF /i "%cancel_script%"=="o" goto:define_action_type
+set daybreak_method_choice=
 IF "%action_type%"=="1" (
 	call "%associed_language_script%" "copying_begin"
 	if exist "%volume_letter%:\FW_%firmware_choice%" rmdir /s /q "%volume_letter%:\FW_%firmware_choice%"
 	%windir%\System32\Robocopy.exe "firmware_temp " %volume_letter%:\FW_%firmware_choice% /e >nul
-	call :daybreak_convert "%volume_letter%:\FW_%firmware_choice%"
+	echo.
+	call "%associed_language_script%" "daybreak_convert_choice"
+	IF "!daybreak_method_choice!"=="1" (
+		call :daybreak_convert "%volume_letter%:\FW_%firmware_choice%"
+	)
+	IF "!daybreak_method_choice!"=="2" (
+		call :daybreak_convert_2 "%volume_letter%:\FW_%firmware_choice%"
+	)
 	IF EXIST "%volume_letter%:\switch\ChoiDuJourNX.nro" del /q "%volume_letter%:\switch\ChoiDuJourNX.nro"
 	IF NOT EXIST "%volume_letter%:\switch" mkdir "%volume_letter%:\switch"
 	IF NOT EXIST "%volume_letter%:\switch\ChoiDuJourNX" mkdir "%volume_letter%:\switch\ChoiDuJourNX"
@@ -758,7 +766,11 @@ set keys_file_path=
 call "%associed_language_script%" "daybreak_keys_file_select"
 IF "%keys_file_path%"=="" (
 	call "%associed_language_script%" "daybreak_keys_file_select_passed"
-	pause
+	IF NOT "!temp_choice!"=="" set temp_choice=!temp_choice:~0,1!
+	call "tools\Storage\functions\modify_yes_no_always_never_vars.bat" "temp_choice" "o/n_choice"
+	IF /i "!temp_choice!"=="O" (
+		call :daybreak_convert_2 "%~1"
+	)
 	exit /b
 )
 call "%associed_language_script%" "daybreak_convert_begin"
@@ -779,7 +791,11 @@ for %%f in ("%~1\*.nca") do (
 			set /p temp_count=<templogs\tempvar.txt
 			IF NOT "!temp_count!"=="0" (
 				call "%associed_language_script%" "daybreak_convert_keys_warning"
-				pause
+				IF NOT "!temp_choice!"=="" set temp_choice=!temp_choice:~0,1!
+				call "tools\Storage\functions\modify_yes_no_always_never_vars.bat" "temp_choice" "o/n_choice"
+				IF /i "!temp_choice!"=="O" (
+					call :daybreak_convert_2 "%~1"
+				)
 				exit /b
 			)
 			rem pause
@@ -798,7 +814,11 @@ for %%f in ("%~1\*.nca") do (
 			set /p temp_count=<templogs\tempvar.txt
 			IF NOT "!temp_count!"=="0" (
 				call "%associed_language_script%" "daybreak_convert_keys_warning"
-				pause
+				IF NOT "!temp_choice!"=="" set temp_choice=!temp_choice:~0,1!
+				call "tools\Storage\functions\modify_yes_no_always_never_vars.bat" "temp_choice" "o/n_choice"
+				IF /i "!temp_choice!"=="O" (
+					call :daybreak_convert_2 "%~1"
+				)
 				exit /b
 			)
 			rem pause
@@ -808,6 +828,36 @@ for %%f in ("%~1\*.nca") do (
 			IF "!temp_count!"=="0" (
 			rem echo NCA mal nommé.
 			move "!temp_file_name!" "!temp_file_name:~0,-8!nca" >nul
+			)
+		)
+	)
+)
+rem echo %count_loop%
+exit /b
+
+:daybreak_convert_2
+rem set /a count_loop = 0
+for /d %%f in ("%~1\*.nca") do (
+	move "%%f" "%%f.bak" >nul
+	move "%%f.bak\00" "%%f" >nul
+	rmdir "%%f.bak"
+)
+for %%f in ("%~1\*.nca") do (
+	set temp_file_size=%%~zf
+	IF NOT EXIST "%%f\*.*" (
+		set temp_file_name=%%f
+		IF NOT "!temp_file_name:~-9,9!"==".cnmt.nca" (
+			rem set /a count_loop = !count_loop!+1
+			IF !temp_file_size! LEQ 8192 (
+				rem echo Meta trouvé.
+				move "!temp_file_name!" "!temp_file_name:~0,-3!cnmt.nca" >nul
+			)
+		) else IF "!temp_file_name:~-9,9!"==".cnmt.nca" (
+			rem set /a count_loop = !count_loop!+1
+			IF NOT !temp_file_size! LEQ 8192 (
+				rem echo %%f
+				rem echo NCA mal nommé.
+				move "!temp_file_name!" "!temp_file_name:~0,-8!nca" >nul
 			)
 		)
 	)
